Broke down index page forms to smaller components
