+++ app/types.ts
@@ -1,5 +1,6 @@
 import ReactPlayer from 'react-player';
 import { MutableRefObject } from 'react';
+import { UseFormReturnType } from '@mantine/form';
 
 export interface AppContextType {
   username: string;
@@ -31,3 +32,19 @@ export interface Player {
   playbackRate: number;
   reactPlayerRef?: MutableRefObject<ReactPlayer | null>;
 }
+
+export type FormRef<T> = UseFormReturnType<T, (values: T) => T>;
+
+export interface ExistingRoomFormValues {
+  username: string;
+  url: string;
+}
+
+export interface CreateRoomFormValues {
+  username: string;
+  videoUrl: string;
+}
+
+export type ExistingRoomFormRef = FormRef<ExistingRoomFormValues>;
+
+export type CreateRoomFormRef = FormRef<CreateRoomFormValues>;
